requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.8.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.25
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
Flask==3.0.2
Flask-SQLAlchemy==2.5.1
future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gspread==5.12.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.1.1
pillow==10.2.0
psycopg2==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
PyJWT==2.1.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.9
simple-term-menu==1.6.3
six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.4.46
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.6
Werkzeug==3.0.1
zipp==3.17.0
